Branch MatrixPilot_UDB5_WJP: new project MatrixPilot-auav3_noFS_MDD_USB:  reverse previous commit and change barometer output to floating point pressure altitude in the vfr_hud record

+++ libDCM/estAltitude.h
@@ -27,7 +27,7 @@ void udb_barometer_callback(long pressure, int temperature, char status);
 void altimeter_calibrate(void);
 void estAltitude(void);
 
-inline long get_barometer_altitude(void);
+inline float get_barometer_altitude(void);
 inline long get_barometer_pressure(void);
 inline int get_barometer_temperature(void);
 inline long get_barometer_agl_altitude(void);
